Refactor: add Remote data source for favourite cocktail
