--- app/src/main/java/com/sun/cocktaildb/screen/cocktaildetail/CocktailDetailPresenter.kt
@@ -1,19 +1,0 @@
-package com.sun.cocktaildb.screen.cocktaildetail
-
-import com.sun.cocktaildb.utils.base.BasePresenter
-
-class CocktailDetailPresenter : BasePresenter<CocktailDetailView> {
-    private var view: CocktailDetailView? = null
-
-    override fun setView(view: CocktailDetailView?) {
-        this.view = view
-    }
-
-    override fun onStart() {
-        // No-op for initial skeleton
-    }
-
-    override fun onStop() {
-        // No-op for initial skeleton
-    }
-}
