--- app/src/main/java/com/sun/cocktaildb/screen/cocktaildetail/CocktailDetailView.kt
@@ -1,5 +1,0 @@
-package com.sun.cocktaildb.screen.cocktaildetail
-
-import com.sun.cocktaildb.utils.base.BaseView
-
-interface CocktailDetailView : BaseView
